requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.88.0
youtube-transcript-api>=0.6.0
openai>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0 